tchannel/http: move handler into a separate file
